pyproject.toml
[project]
name = "python-django-diploma"
version = "0.1.0"
description = ""
authors = [
    {name = "Дмитрий Видякин",email = "vidyakindy@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-stubs (>=3.16.8,<4.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "setuptools (>=82.0.0,<83.0.0)",
    "django-filter (>=25.2,<26.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "whitenoise (>=6.12.0,<7.0.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
